Refactor data fetching to a hook
